rename it to keep unify with other component
